# docs/swagger.yaml
definitions:
...
  dtos.ErrorIndicator:
    properties:
      ErrorId:
        $ref: '#/definitions/entities.ErrorType'
      count:
        type: integer
      errorType:
        type: string
    type: object
  dtos.ErrorResponse:
    properties:
      message:
        type: string
      timestamp:
        type: string
    type: object
  dtos.ErrorsCountPerDayAndType:
    properties:
      date:
        type: string
      errorType:
        type: string
      total:
        type: integer
    type: object
...
  entities.ErrorType:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - NullValue
    - Missing4DaysHorizon
    - OutsideOfErrorBands
    - ForecastMissing
...
info:
  contact: {}
  description: This API manages the inputs of the forecast model
  title: Inputs API
  version: "1.0"
paths:
...
  /errores/indicadores:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: hoy'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: Id de la configuracion
        format: uint
        in: query
        name: configurationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.ErrorIndicator'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener las indicadores de errores
      tags:
      - Errores
  /errores/por-dia:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: hoy'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: ID de la configuracion
        in: query
        name: configurationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.ErrorsCountPerDayAndType'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener las errores detectados por dia
      tags:
      - Errores
  /errores/series-implicadas:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: hoy'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: Id de la configuracion
        format: uint
        in: query
        name: configurationId
        required: true
        type: string
      - description: Id del tipo de error
        format: uint
        in: query
        name: errorType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.ErrorIndicator'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener las indicadores de errores
      tags:
      - Errores
...
swagger: "2.0"